+++ src/main/java/br/com/zalf/prolog/webservice/autenticacao/AutenticacaoDao.java
@@ -1,10 +1,10 @@
 package br.com.zalf.prolog.webservice.autenticacao;
 
-import br.com.zalf.prolog.webservice.interceptors.auth.authenticator.StatusSecured;
+import br.com.zalf.prolog.webservice.interceptors.auth.ColaboradorAutenticado;
 import org.jetbrains.annotations.NotNull;
 
-import java.sql.SQLException;
 import java.time.LocalDate;
+import java.util.Optional;
 
 /**
  * Autenticação do usuário no sistema.
@@ -14,22 +14,21 @@ public interface AutenticacaoDao {
      * Cria um novo token para o usuário. Esse Token será utilizado para fazer a validação
      * do usuário que está requisitando informações do servidor.
      *
-     * @param cpf Cpf do usuário.
+     * @param cpf CPF do usuário.
      * @return Objeto {@link Autenticacao autenticação} com o token gerado e o status do login.
-     * @throws SQLException Caso não seja possível criar o token na tabela.
+     * @throws Throwable Caso ocorra algum erro.
      */
     @NotNull
-    Autenticacao insertOrUpdate(@NotNull final Long cpf) throws SQLException;
+    Autenticacao insertOrUpdate(@NotNull final Long cpf) throws Throwable;
 
     /**
-     * Deleta um token da tabela. Este método é usado quando o usuário
-     * solicita logout do sistema.
+     * Deleta um token da tabela. Este método é usado quando o usuário realiza logout do sistema.
      *
-     * @param token Contém o CPF e o token a ser deletado da tabela.
+     * @param token Token a ser deletado da tabela.
      * @return Resultado da operação, true ou false.
-     * @throws SQLException Caso não seja possível fazer a deleção do token na tabela.
+     * @throws Throwable Caso ocorra algum erro.
      */
-    boolean delete(@NotNull final String token) throws SQLException;
+    boolean delete(@NotNull final String token) throws Throwable;
 
     /**
      * Verifica a existência de um token.
@@ -37,10 +36,12 @@ public interface AutenticacaoDao {
      * @param token                O token que queremos verificar.
      * @param apenasUsuariosAtivos Indica se devemos considerar na verificação apenas usuário
      *                             que estão ativados no sistema (STATUS_ATIVO = true).
-     * @return Boolean com o resultado da requisição.
-     * @throws SQLException Caso não seja possível realizar a busca.
+     * @return {@link Optional optional} que irá conter o colaborador autenticado caso ele exista no sistema
+     * @throws Throwable Caso ocorra algum erro.
      */
-    boolean verifyIfTokenExists(@NotNull final String token, final boolean apenasUsuariosAtivos) throws SQLException;
+    @NotNull
+    Optional<ColaboradorAutenticado> verifyIfTokenExists(@NotNull final String token,
+                                                         final boolean apenasUsuariosAtivos) throws Throwable;
 
     /**
      * Verifica a existência de um CPF e data de nascimento na base de dados.
@@ -49,12 +50,13 @@ public interface AutenticacaoDao {
      * @param dataNascimento       Data de Nascimento do colaborador a ser verificada a existência.
      * @param apenasUsuariosAtivos Indica se devemos considerar na verificação apenas usuário
      *                             que estão ativados no sistema (STATUS_ATIVO = true).
-     * @return Valor booleano que representa se o usuário está cadastrado no banco de dados.
-     * @throws SQLException Caso não seja possível verificar a existência no banco de dados.
+     * @return {@link Optional optional} que irá conter o colaborador autenticado caso ele exista no sistema.
+     * @throws Throwable Caso ocorra algum erro.
      */
-    boolean verifyIfUserExists(@NotNull final Long cpf,
-                               @NotNull final LocalDate dataNascimento,
-                               final boolean apenasUsuariosAtivos) throws SQLException;
+    @NotNull
+    Optional<ColaboradorAutenticado> verifyIfUserExists(@NotNull final Long cpf,
+                                                        @NotNull final LocalDate dataNascimento,
+                                                        final boolean apenasUsuariosAtivos) throws Throwable;
 
     /**
      * Verifica se o usuário tem as permissões necessárias para acessar determinada função.
@@ -65,14 +67,14 @@ public interface AutenticacaoDao {
      *                                  as permissões passadas no array no mínimo uma.
      * @param apenasUsuariosAtivos      Indica se devemos considerar na verificação apenas usuário
      *                                  que estão ativados no sistema (STATUS_ATIVO = true).
-     * @return Objeto {@link StatusSecured StatusSecured} com o estado atual da validação do token e permissões.
+     * @return {@link Optional optional} que irá conter o colaborador autenticado caso ele exista no sistema
      * @throws Throwable Caso ocorra algum erro.
      */
     @NotNull
-    StatusSecured userHasPermission(@NotNull final String token,
-                                    @NotNull final int[] permissions,
-                                    final boolean needsToHaveAllPermissions,
-                                    final boolean apenasUsuariosAtivos) throws Throwable;
+    Optional<ColaboradorAutenticado> userHasPermission(@NotNull final String token,
+                                                       @NotNull final int[] permissions,
+                                                       final boolean needsToHaveAllPermissions,
+                                                       final boolean apenasUsuariosAtivos) throws Throwable;
 
     /**
      * Verifica se o usuário tem as permissões necessárias para acessar determinada função.
@@ -84,13 +86,13 @@ public interface AutenticacaoDao {
      *                                  as permissões passadas no array ou mínimo uma.
      * @param apenasUsuariosAtivos      Indica se devemos considerar na verificação apenas usuário
      *                                  que estão ativados no sistema (STATUS_ATIVO = true).
-     * @return Objeto {@link StatusSecured StatusSecured} com o estado atual da validação do token e permissões.
+     * @return {@link Optional optional} que irá conter o colaborador autenticado caso ele exista no sistema
      * @throws Throwable Caso ocorra algum erro.
      */
     @NotNull
-    StatusSecured userHasPermission(final long cpf,
-                                    @NotNull final LocalDate dataNascimento,
-                                    @NotNull int[] permissions,
-                                    final boolean needsToHaveAllPermissions,
-                                    final boolean apenasUsuariosAtivos) throws Throwable;
+    Optional<ColaboradorAutenticado> userHasPermission(final long cpf,
+                                                       @NotNull final LocalDate dataNascimento,
+                                                       @NotNull int[] permissions,
+                                                       final boolean needsToHaveAllPermissions,
+                                                       final boolean apenasUsuariosAtivos) throws Throwable;
 }