+++ src/main/java/br/com/zalf/prolog/webservice/integracao/response/PosicaoPneuMepadoResponse.java
@@ -1,0 +1,40 @@
+package br.com.zalf.prolog.webservice.integracao.response;
+
+import lombok.Data;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+/**
+ * Created on 31/03/20.
+ *
+ * @author Natan Rotta (https://github.com/natanrotta)
+ */
+@Data
+public class PosicaoPneuMepadoResponse {
+    @NotNull
+    public static final String GENERIC_ERROR_MESSAGE = "Não foi possível cadastrar o pneu no Sistema ProLog";
+    private final boolean success;
+    private final int codDiagrama;
+    @Nullable
+    private final String errors;
+    @Nullable
+    private final Throwable throwable;
+
+    @NotNull
+    public static PosicaoPneuMepadoResponse ok(final int codDiagrama) {
+        return new PosicaoPneuMepadoResponse(true, codDiagrama, null, null);
+    }
+
+    @NotNull
+    public static PosicaoPneuMepadoResponse error(final int codDiagrama,
+                                                  @NotNull final String errors) {
+        return new PosicaoPneuMepadoResponse(false, codDiagrama, errors, null);
+    }
+
+    @NotNull
+    public static PosicaoPneuMepadoResponse error(final int codDiagrama,
+                                                  @NotNull final String errors,
+                                                  @NotNull final Throwable throwable) {
+        return new PosicaoPneuMepadoResponse(false, codDiagrama, errors, throwable);
+    }
+}