+++ src/main/java/br/com/zalf/prolog/webservice/interceptors/debugenv/ResourceDebugOnlyInterceptor.java
@@ -1,6 +1,6 @@
 package br.com.zalf.prolog.webservice.interceptors.debugenv;
 
-import br.com.zalf.prolog.webservice.BuildConfig;
+import br.com.zalf.prolog.webservice.config.BuildConfig;
 
 import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.container.ContainerRequestFilter;